Transceiver double change

+++ src/main/java/net/povstalec/sgjourney/common/compatibility/cctweaked/methods/TransceiverMethods.java
@@ -21,7 +21,7 @@ public class TransceiverMethods
 		@Override
 		public MethodResult use(IComputerAccess computer, ILuaContext context, TransceiverEntity transceiver, IArguments arguments) throws LuaException
 		{
-			long frequency = arguments.getLong(0);
+			double frequency = arguments.getDouble(0);
 			if(frequency > Integer.MAX_VALUE || frequency < Integer.MIN_VALUE)
 				throw new LuaException("Frequency " + frequency + " out of range for <" + Integer.MIN_VALUE + ", " + Integer.MAX_VALUE + ">");
 			TransceiverFunctions.setFrequency(transceiver, (int) frequency);
